% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk_detect.R
\name{tk_detect}
\alias{tk_detect}
\title{Detect Thermokarst from Elevation}
\usage{
tk_detect(elev, radii = 15, cutoff = 0, n.cores = 1)
}
\arguments{
\item{n.cores}{}
}
\value{

}
\description{
Detect thermokarst features by comparing the elevation in each cell of a
digital terrain model to the median elevation in a circular neighborhood.
}
