% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk_ponds.R
\docType{data}
\name{tk_ponds}
\alias{tk_ponds}
\title{Elevation raster (DTM) of some thermokarst pits.}
\format{
A RasterBrick.
}
\source{
\url{https://data.neonscience.org/data-products/DP3.30024.001}
}
\usage{
tk_ponds
}
\description{
A raster brick of the ground surface elevation of some thermokarst ponds
where the two layers are different years of data.
}
\keyword{datasets}
